rename and move PolicyHandler file to types folder
