+++ src/auth/types/policy-handler.type.ts
@@ -1,4 +1,4 @@
-import { AppAbility } from './casl/casl-ability.factory';
+import { AppAbility } from '../casl/casl-ability.factory';
 
 interface IPolicyHandler {
   handle(ability: AppAbility): boolean;
